+++ semcore/feedback-form/src/component/submit-button/SubmitButton.tsx
@@ -1,0 +1,9 @@
+import Button from '@semcore/button';
+import React from 'react';
+import { Root, sstyled, IRootComponentProps } from '@semcore/core';
+
+export function SubmitButton(props: IRootComponentProps) {
+  const { styles } = props;
+  const SSubmit = Root;
+  return sstyled(styles)(<SSubmit render={Button} type='submit' use='primary' theme='success' />);
+}