+++ semcore/flags/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.25.0] - 2024-05-21
+
+### Removed
+
+- `aria-label` from the flag component as it was duplicating country name commonly placed near the flag.
+
+### Added
+
+- `FlagsIso2` and `FlagsIso3` types export.
+
 ## [4.24.1] - 2024-05-17
 
 ### Changed